lib/book/requirements.txt
ark
pybtex
pymdown-extensions
python-frontmatter
ruff
